Place the forwarded ssh-agent socket to the runtime directory

+++ session.c
@@ -170,6 +170,7 @@ static char *auth_info_file = NULL;
 /* Name and directory of socket for authentication agent forwarding. */
 static char *auth_sock_name = NULL;
 static char *auth_sock_dir = NULL;
+int remove_auth_sock_dir = 1;
 
 /* removes the agent forwarding socket */
 
@@ -179,8 +180,13 @@ auth_sock_cleanup_proc(struct passwd *pw)
 	if (auth_sock_name != NULL) {
 		temporarily_use_uid(pw);
 		unlink(auth_sock_name);
-		rmdir(auth_sock_dir);
+		free(auth_sock_name);
 		auth_sock_name = NULL;
+
+		if (remove_auth_sock_dir)
+			rmdir(auth_sock_dir);
+		free(auth_sock_dir);
+		auth_sock_dir = NULL;
 		restore_uid();
 	}
 }
@@ -189,6 +195,7 @@ static int
 auth_input_request_forwarding(struct ssh *ssh, struct passwd * pw)
 {
 	Channel *nc;
+	char *path;
 	int sock = -1;
 
 	if (auth_sock_name != NULL) {
@@ -199,17 +206,25 @@ auth_input_request_forwarding(struct ssh *ssh, struct passwd * pw)
 	/* Temporarily drop privileged uid for mkdir/bind. */
 	temporarily_use_uid(pw);
 
-	/* Allocate a buffer for the socket name, and format the name. */
-	auth_sock_dir = xstrdup("/tmp/ssh-XXXXXXXXXX");
+	path = session_get_runtime_directory();
+	if (strcmp(path, "/tmp") == 0) {
+		/* Allocate a buffer for the socket name, and format the name. */
+		auth_sock_dir = xstrdup("/tmp/ssh-XXXXXXXXXX");
 
-	/* Create private directory for socket */
-	if (mkdtemp(auth_sock_dir) == NULL) {
-		packet_send_debug("Agent forwarding disabled: "
-		    "mkdtemp() failed: %.100s", strerror(errno));
-		restore_uid();
-		free(auth_sock_dir);
-		auth_sock_dir = NULL;
-		goto authsock_err;
+		/* Create private directory for socket */
+		if (mkdtemp(auth_sock_dir) == NULL) {
+			packet_send_debug("Agent forwarding disabled: "
+			    "mkdtemp() failed: %.100s", strerror(errno));
+			restore_uid();
+			free(auth_sock_dir);
+			auth_sock_dir = NULL;
+			goto authsock_err;
+		}
+		free(path);
+	} else {
+		/* This is already private directory */
+		auth_sock_dir = path;
+		remove_auth_sock_dir = 0;
 	}
 
 	xasprintf(&auth_sock_name, "%s/agent.%ld",
@@ -236,7 +251,8 @@ auth_input_request_forwarding(struct ssh *ssh, struct passwd * pw)
  authsock_err:
 	free(auth_sock_name);
 	if (auth_sock_dir != NULL) {
-		rmdir(auth_sock_dir);
+		if (remove_auth_sock_dir)
+			rmdir(auth_sock_dir);
 		free(auth_sock_dir);
 	}
 	if (sock != -1)
